requirements.txt
torch>=2.2
transformers>=4.41
accelerate>=0.31
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
opencv-python
Pillow
tqdm
matplotlib
seaborn
huggingface_hub>=0.23
gradio>=4.36
requests
ujson
